+++ CHANGELOG.md
@@ -4,7 +4,8 @@ CHANGELOG
 Next release
 ------------
 
-* Add few behaviors for dom events (copy/confirm).
+* Add Sentry helpers.
+* Add few behaviors for dom events (copy/confirm/processing).
 * Add helpers for work with attributes and states of elements.
 * Add UI notifications with adapters (toastr, notyf).